requirements.txt
faiss_cpu==1.10.0
langchain_core==0.3.49
langchain_groq==0.3.1
numpy==1.21.5
PyGithub==2.6.1
PyGithub==2.6.1
python-dotenv==1.1.0
Requests==2.32.3
sentence_transformers==3.4.1
